Adding type hints on variable, class variable, and functions
Summary:
This patch adds type hints on variables, class variables and as
return type of functions in the PHP grammar as well as pfff AST.
It also add an action in ppp, --unparse_without_type_hints which
given a file parse the file and remove all the type annotations
(so that it can be parsed by a vanilla PHP parser such as the one
in HPHP or Zend).

Test Plan:
./ppp -unparse_without_type_hints tests/php/typehints/foo.php
will correctly parse and unparse a file on stdout
where all annotations have been removed.

I've also created a ~pad/bin/ppp so you can play with it.

DiffCamp Revision: 168394
Reviewed By: asuhan
CC: pad, asuhan
Revert Plan:
OK