pfff: introduce self/parent tokens in lexer (useful for static:: handling later)
Summary:
The "self" and "parent" keywords were actually not considered
as keywords in the original Zend PHP lexer (which I mostly copy paste
in pfff). Now that static:: can be used in many places where we
now use self:: or parent::, and since static is actually a keyword
(recognized as such by the Zend lexer), this patch refactors the lexer
and AST to better reflect those notions.

Before this patch self/parent were considered as IDENT and recognized
manually in a few places in the grammar and post-grammar analysis. I
was for instance recognizing the "self"/"parent" strings when used in
a qualifier position as in self::. But I overlooked a few other places
where one can use the self/parent sugar as in 'new self()'. Having "self"
and "parent" as tokens generate parse errors in those situations which
is a good thing because it forces me to extend the grammar and AST
and to better express what PHP is.

This patch just includes the modifications to the lexer, grammar and AST.
There are lots of collateral evolutions to the change of type
(including in the code of checkModuleFast),
but they are mostly trivial and flagged automatically by
the ocaml compiler  so I have not included them in this patch.

Task ID: #

Blame Rev:

Reviewers: eletuchy, bill, erling

CC: mathewsb

Test Plan:
$ ./pfff -parse_php tests/php/parsing/
does not generate parse error on the new test files.
$ ./pfff_test all
no regression

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 238447