Pfff/tags: generate :x:foo and also x:foo tags for xhp classes
Summary:
Paul wanted that. It can make things easier for Emacs or Vim
as the default functions to get the ident under the cursor
will not magically prepend a : for xhp classes.

As paul said: "I like to select the line where the XHP
call is being done and jump to it. Adding a colon in front sucks."

Test Plan:
unit test included.
$ make test

Reviewers: julienv

Reviewed By: julienv

CC: ptarjan, yidingjia, ashoat, eletuchy, julienv

Differential Revision: https://phabricator.fb.com/D387310

Task ID: 855568