Array guessing tool
Summary: Added more evidence to assist in guessing each array, as well as corrected the definition of a tuple. Evidence used now includes bracket access of an array ($arr[]), array declaration types, the types of values added to the array, and the types of the keys of the array.

Test Plan: Added unit tests in unit_array_inference_php.ml, but need to flush out and add more.

Reviewers: julienv, pad

CC:

Task ID: #1146337

Blame Rev: