Fixed tuple patch and print out code snippet of evidence
Summary: Fixed the tuple patch in order to print out a tuple of the types associated with a tuple used as a parameter or a return value. Also added error handling for containers that can not be typed, ie a vector of an unknown type etc.When displaying evidence, a line of code is now displayed.

Test Plan:

Reviewers: julienv

CC:

Task ID: #1146337

Blame Rev: