add json output mode for cyclomatic information
Summary:
To aid in some generating the "CRAP" number for code
(http://www.crap4j.org/), the test runner needs to collect the
cyclomatic complexity in addition to the coverage metrics.

This diff adds a simple JSON output mode using the json library that is
already present in the pfff repo.

Test Plan:
1. produce a php file that contains functions and/or methods
2. ./facebook/qa_test/qa_test -print_cyclomatic_json file.php
3. inspect the output; it should be a JSON object whose keys are the
path(s) to the files being inspected, and whose values are arrays of
objects containing the function name, complexity number and the starting
line of the function definition:

$ ./facebook/qa_test/qa_test -print_cyclomatic_json ~/foo.php ~/local/www-git/zeusgodofthunder/index.php
{
 "/home/wez/foo.php": [
   { "func": "otherfunc", "comp": 1, "line": 31 },
   { "func": "func", "comp": 2, "line": 23 }
 ],
 "/home/wez/local/www-git/zeusgodofthunder/index.php": []
}

Reviewers: pad

Reviewed By: pad

Differential Revision: https://phabricator.fb.com/D536888

Task ID: 1292316