Manual patch for confused arrays
Summary: Added a prompt to manually select the container and enter the type for confused arrays. Also fixed a bug so that key and values with no type information do not negatively impact inferencing algorithm. Basically ignore the type if it can not be inferred, in order to not save the uninferred type, Tsum[], as the type of the key or value.

Test Plan: Manual tests, need to add unit tests

Reviewers: julienv

Differential Revision: https://phabricator.fb.com/D537082

Task ID: 1146337