prolog: index constant arguments to newv, DT, via a docall(..., special)
Summary:
Mergen needs that.
He wants to find all call sites that have "DT(X)->add" where
XDataType is a subclass of SomethingDataType.
Unfortunately the callgraph in the prolog database index
only simple calls and does not index the arguments to special functions
like DT or newv. This patch fixes that.

An alternative would be to combine sgrep with cmf --prolog, but
I think indexing wrappers to new are a cleaner solution.

Test Plan:
unit test included
$ make test

Reviewers: pieter, julienv, bill, amenghra, michalburger1

Reviewed By: pieter

CC: mnachin, benjaminstrahs, ahupp, eletuchy, kbeck, abrady, pdimov, nharper, mroch, jlindamood, mkedia, lovro, erling, ptarjan, mathieubaudet

Differential Revision: https://phabricator.fb.com/D600162