cmf: optimize loading autoload file, part1
Summary:
It currently takes 7.4s to cmf to parse the autoload file
(see cmf -profile). This patch introduces a new parsing
function for PHP which removes many features that are not
necessary for what we do on the autoload file.
It now takes 1.5s to parse the huge autoload file.

There are other optimizations we could do:
 - cache the parsed AST on the disk
 - not use the PHP parser at all and make a parsing function
   specialized for the autoload file
 - use julien's fbstrict parser (dunno if faster on such a file)
 - use hiphop parser
 - ...

The next patch will modify check_module to use this new function.

Test Plan:
$ time ~/pfff/pfff.opt -profile -parse_php_fast autoload_map.php
1.5s

Reviewers: pieter, zoel

Reviewed By: pieter

CC: julienv, andrewparoski, jakubv, ptarjan

Differential Revision: https://phabricator.fb.com/D658432

Task ID: 1915724