java: generate inheritance information in prolog format
Summary:
Now that we can parse the .java and .class files in fb4a, and that we have
the call graph, inheritance tree, and data graph (without any errors),
we can start generate prolog facts about the codebase to make it easy
to write certain analysis. With this patch one can easily get
the list of parents, recursively, or children, of a class.

The next patches will encode the call graph in a similar format,
as well as the data graph, which will then make it easy to ask questions
such as: is there any method calling startActivity inside a class
that derives from Fragment.

Test Plan:
unit test included
$ cd pfff
$ make
$ make test

Reviewers: pieter

Reviewed By: pieter

CC: bill, herrys, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D674744

Task ID: 1779304