java: generate call graph information in prolog format
Summary:
This patch makes accessible the call graph in prolog.

It's based on the bytecode (.class files), which is typed, which makes it
trivial to resolve complex method chain calls (no need
for type inference, which is needed for analyzing recent .java files).

Note that the call graph is nevertheless imprecise as
a call to a method (Foo,bar) can in reality be a call
to a derived class of Foo. To be more precise requires
global analysis, such as RTA, XTA, context sensitive interprocedural
analysis, object sensitive interprocedural analysis, etc.

RTA is relativly easy to implement, so I'll probably do that next,
(if it's actually needed by some of our bug analysis).

Test Plan:
unit tests included
$ make tests

Reviewers: pieter

Reviewed By: pieter

CC: bill, herrys, sidd, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D674816

Task ID: 1779304