Add type arguments to the PFFF ast
Summary:
This adds type arguments to the PFFF ast.

There are a bunch of TODOs; basically I'm assuming that if I don't
recognize a particular module, it's probably not actively deployed
somewhere (and it is safe to ignore type args). Feel free to comment
if you think some of them should be addressed now.

Test Plan: Ran `make test` with 125 tests; 5 skips.

Reviewers: pad

Reviewed By: pad

CC: amenghra, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D735084