Add AssignInBooleanContext lint rule
Summary:
Port the lint check that catches assignments in boolean context
from cmf to pfff so it'll get executed as part of scheck.

This uses the formerly disabled "strict" version of the check as
the non-strict one was broken in that it wasn't possible to silence
the warning by adding parentheses around the assignment expression.

Test Plan:
* added test cases for the new code
* fixed "unexpectedly succeeding" test case
* ran tests (./pfff_test checkers_php)

Reviewers: zoel, pad

Reviewed By: pad

CC: ericsumner

Differential Revision: https://phabricator.fb.com/D744421

Task ID: 2172813