prolog: handling the new PHP syntax for (new Xxx())->foo()
Summary:
Jyan detected bugs in cmf --prolog where some calls to new
were not present in the prolog DB. It's because of the
new construct in ast_php.ml which was not take into account
in the prolog fact generator. This patch fixes that.

thx jyan.

Test Plan:
unit test included
$ make test

Reviewers: bill, pieter, zoel, jyan

Reviewed By: pieter

CC: ptarjan, julienv, amenghra

Differential Revision: https://phabricator.fb.com/D763650