pfff/php: big cleanup of the grammar part1
Summary:
There are currently 3 parsing errors in www on code like foo()[2]->method().
The pfff grammar has 7 shift/reduce conflicts. The grammar
has gradually become ugly and hard to extend. Same for the AST.
I used to follow exactly the Zend grammar,
even when it looked stupid because of many special cases, but now,
with the many extensions we've added, it's time to rewrite the grammar
(and AST) with more orthogonal rules which by side effect should avoid
the many parsing errors and special cases.

The next diff will attack the expr/lvalue grammar rules, which
are the messiest, but for now I wanted to cleanup all the rest.

- removed some useless intermediate rules
- removed %start rules not used
- factorized trait_declaration_statement with class_declaration_statement
- explain in conflicts.txt impossibility to factorize things withtype_opt
- ext_type_hint -> type_php and simplification of type_php
- factorize some rules
- ident vs qualified vs name
- many more stuff

Btw here are some stats about our different PHP parsers:

  pfff:
  180 terminals, 191 nonterminals
  625 grammar rules, 1118 states
  7 s/r conflicts

  hack:
  132 terminals, 83 nonterminals
  329 grammar rules, 663 states
  3 s/r conflicts

  hphp:
  ? terminals, ? nonterminals
  710 grammar rules, 1329 states
  2 s/r conflicts

  xhpast:
  ? terminals, ? nonterminals
  584 grammar rules, 1144 states
  12 s/r conflicts

Test Plan: make test

Reviewers: pieter

CC: julienv, amenghra, jdelong

Differential Revision: https://phabricator.fb.com/D815823