pfff: handle the trait renaming feature in the AST
Summary:
Every week I got some emails because 'arc lint' returns a parse
error when people try to use the renaming feature of traits.
Then I have to tell them this feature is not supported
because it leads to very hard to follow code.

To avoid those emails it's better to have a better
error message in arc lint, but the first step is
to actually represent the feature in the AST.
The next diff will modify cmf to print the right
error message.

Test Plan:
unit test included
$ make test

Reviewers: pieter, jiaol, mathieubaudet, zoel

CC: julienv

Differential Revision: https://phabricator.fb.com/D845100