codegraph/php: remove false positives on "lookup failure method prepareXxx"
Summary:
Jiao found this false positive in the log of codegraph. This is because
of the very dynamic "DynamicYield" trait. For now we want to
focus on real bugs in www and fix all of them gradually so let's
skip error related to prepareXxxx for now.

Test Plan: make test

Reviewers: jiaol, pieter, mathieubaudet

Reviewed By: jiaol

CC: glevi, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D850067