codegraph/php: dont confuse ClassConstant and Fields
Summary:
Jiao found some bugs in the log of codegraph where we were reporting
the use of undefined ClassConstant whereas it was actually a Field.
This is because of some refactoring I've done 2 weeks ago where I
splitted Id in Id and Var in ast_php_simple.

Test Plan: make test

Reviewers: jiaol, pieter, mathieubaudet

Reviewed By: jiaol

CC: platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D850583