pfff: support for namespace part5, collateral evolutions on code using the AST
Summary:
Now that a name can be a list of string, the code using the AST
and mentioning names have to be updated. Again most of the code
was updated via spatch. The rest was modified manually when
the typechecker pinpointed all the remaining places to fix.

Test Plan:
make
it compiles again!

Reviewers: pieter, julienv, jiaol, amenghra, glevi, zoel

Reviewed By: pieter

CC: jdelong, ptarjan, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D883200