codegraph: handle newtype and type, part1
Summary:
The pfff cron is currently failing because some code in
www is using the new type and newtype declarations.

Codegraph used to incorrectly visit the type value
in the Defs phase which lead to some lookup failure because
once need first to extract all the class definitions
before trying to extract the uses (in the second Uses phase).

Test Plan:
unit test included
make graph does not generate anymore lookup failure
on the 'B' class.

Reviewers: pieter

Reviewed By: pieter

CC: platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D967942