[abstract interpreter] re-enable foreach stuff
Summary:
Similar to D1262073: minor adjustments because the shape of
ast_php_simple has changed since this was last in production.

Add an explicit exception type that gets thrown when we see
a foreach with a shape we don't (yet?) handle.

Test Plan: Will add corresponding integration tests.

Reviewers: pad, julienv

Reviewed By: pad

CC: collingreene

Differential Revision: https://phabricator.fb.com/D1268141

Task ID: 4081629