Partially fix CMT with 4.02
Summary:
I really want to use ocaml 4.02 trunk for Hack, and need pfff
for that. Here are some build fixes for CMT, before I found the flag to
`configure` that disabled it, which is good enough for Hack.

Not sure if you want to take this or not since it's not finished, but
didn't want a morning's work to go completely to waste.

Test Plan:
`make` gets somewhat further than before, before it errors
out.

Reviewers: pad

Reviewed By: pad

Differential Revision: https://phabricator.fb.com/D1283167