[pfff] move commons_graph to a separate directory
Summary:

Looks like 378a73c2793afdd725e7a1669c128c138787fbba wasn't
quite match how the library is actually built. This moves
the commons graph wrapper to a separate directory so that
it's easier to build as a separate findlib package.

This makes it easier to deal with some internal tooling that
expects a single `.cmxa` and `.a` as a limitation.

Test Plan:

  make clean
  make depend
  make
  make opt
  make uninstall-findlib
  make install-findlib

Reviewers: pieter

Reviewed By: pieter

CC:

Task ID: #

Blame Rev: