[pfff] add support for parsing `??`
Summary:
Similar to the existing `?:` support; produces similar AST nodes.

This is a shortcut and actually has the wrong associativity, since `??` is
right-associative and `?:` is left-associative. We can fix that if we run into
issues; for now we really just need to not throw parse errors on PHP 7
features like this.

Test Plan:
This is a quick fix to ensure that we don't break things
as people start using `??` around the codebase.

Reviewers: mqian, hansonw, leegao

Reviewed By: leegao

Subscribers: dominik, hansonw, alexeyp

Differential Revision: https://phabricator.fb.com/D2770864

Signature: t1:2770864:1450414537:44ccdccf5eda9ade67b08d99686c424d891f3c32