Remove `-warn-error +a` in released builds.
Don't use `-warn-error +a` in released source. If you want to be warning-clean, you should activate the flag only in debug/development mode. Here I'm proposing a simple solution that relies on GNUmake features. You might want to do it in the configure script instead.

As it is now, pfff 0.29 breaks on OCaml 4.03.0+beta1 because of this line (and the use of `String.capitalize`, which is now deprecated).